project [incomplete]
as useful as advice not to start coding immediately was, i needed more time coding and debugging. it's most unfortunate that i looked into this project after our finals concluded: i believe i got a working frame to use, but couldn't implement it as well.
of course, this working frame i mention could benefit from some unused technologies as well, especially in the form of inheritances as commented in code as well.